nodetree.lua: Provide before- and after-callback inspection
A leading and/or a trailing ':' in the callback name indicates whether
nodetree should inspect the callback before or after the currently installed
callback function(s) (or LuaTeX's default functions), respectively.

This feature necessitates a large restructuring of the 'luatexbase' support.
For example, nodetree previously failed with an error if a callback function
was installed for the `mlist_to_hlist` callback (because this is an
'exclusive' callback that doesn't allow a second callback function).

Also add two test files.