nodetree.lua: Don't print `before` and `after` callbacks twice
Move callback wrapper functions to a new table `callback_wrappers` and
populate `callbacks` with wrapper functions to call them.  We need this
additional level of redirection because for the luatexbase API, a callback
function doesn't have access to its own ID string – the previous way of
using `get_print_position` incorrectly returned the same 'before' and
'after' values for both the 'before' and 'after' callbacks, and the only way
to make them different is to have different registered callback functions.

Fixes issue #36.